🧹 tests: extract shared TestFixtures and TestDoubles into TestSupport/
Closes #1446 — extract duplicated makeRunner helper to TestFixtures.swift
Closes #1447 — move SpyLabelsService, SpyConfigStore, SpyProxyStore, HookCounter to TestDoubles.swift

Part of #1379 — Audit our unit-tests
