+++ Tests/RunnerBarCoreTests/SaveRunnerEditsUseCaseTests.swift
@@ -5,73 +5,6 @@ import Foundation
 import RunnerBarCore
 import Testing
 
-// MARK: - Test doubles
-
-/// Spy conformance for `RunnerLabelsService`.
-///
-/// Implemented as an `actor` so Swift's compiler enforces isolation without
-/// requiring `@unchecked Sendable`. The `result` and recorded-call properties
-/// are accessed from the test body before/after `execute(...)` — never concurrently.
-actor SpyLabelsService: RunnerLabelsService {
-    private var result: [String]? = []
-    private(set) var callCount = 0
-    private(set) var lastScope: String?
-    private(set) var lastRunnerID: Int?
-    private(set) var lastLabels: [String]?
-
-    /// Configures the value returned by `patch(...)`. Call from the test body before `execute`.
-    func setUp(result: [String]?) { self.result = result }
-
-    func patch(scope: String, runnerID: Int, labels: [String]) async -> [String]? {
-        callCount += 1
-        lastScope = scope
-        lastRunnerID = runnerID
-        lastLabels = labels
-        return result
-    }
-}
-
-/// Spy conformance for `RunnerConfigStoreProtocol`.
-///
-/// Implemented as an `actor` — see `SpyLabelsService` for rationale.
-actor SpyConfigStore: RunnerConfigStoreProtocol {
-    var loadResult: RunnerConfig = RunnerConfig(workFolder: "_work", disableUpdate: false)
-    private var shouldThrowOnSave = false
-    private(set) var saveCalled = false
-    private(set) var savedConfig: RunnerConfig?
-
-    /// Configures whether `save(...)` throws. Call from the test body before `execute`.
-    func setUp(shouldThrowOnSave: Bool) { self.shouldThrowOnSave = shouldThrowOnSave }
-
-    func load(at _: String) async throws -> RunnerConfig { loadResult }
-    func save(_ config: RunnerConfig, at _: String) async throws {
-        if shouldThrowOnSave { throw TestError.saveFailed }
-        saveCalled = true
-        savedConfig = config
-    }
-}
-
-/// Spy conformance for `RunnerProxyStoreProtocol`.
-///
-/// Implemented as an `actor` — see `SpyLabelsService` for rationale.
-actor SpyProxyStore: RunnerProxyStoreProtocol {
-    private var shouldThrowOnSave = false
-    private(set) var saveCalled = false
-    private(set) var savedConfig: RunnerProxyConfig?
-
-    /// Configures whether `save(...)` throws. Call from the test body before `execute`.
-    func setUp(shouldThrowOnSave: Bool) { self.shouldThrowOnSave = shouldThrowOnSave }
-
-    func load(at _: String) async -> RunnerProxyConfig { RunnerProxyConfig() }
-    func save(_ config: RunnerProxyConfig, at _: String) async throws {
-        if shouldThrowOnSave { throw TestError.saveFailed }
-        saveCalled = true
-        savedConfig = config
-    }
-}
-
-enum TestError: Error { case saveFailed }
-
 // MARK: - Helpers
 
 private func makeRunner(
