--- Tests/RunnerBarCoreTests/TestSupport/.gitkeep
@@ -1,2 +1,0 @@
-# This directory will hold shared test fixtures and doubles.
-# See #1446 and #1447.
